detach extension and binding adapter
