+++ Android/app/src/main/java/co/kr/dgsw/searchvoca/di/module/UseCaseModule.kt
@@ -41,5 +41,6 @@ val wordUseCaseModule = module {
 
 val historyUseCase = module {
     single { GetSearchHistoryCountUseCase(get(), get()) }
+
     single { LoadSearchHistoryUseCase(get(), get()) }
 }
