--- Android/app/src/main/java/co/kr/dgsw/searchvoca/ui/setting/SelectFrequencyDialog.kt
@@ -1,11 +1,0 @@
-package co.kr.dgsw.searchvoca.ui.setting
-
-import co.kr.dgsw.searchvoca.ui.dialog.DefaultBottomSheetDialog
-
-//class SelectFrequencyDialog : DefaultBottomSheetDialog(
-//    listOf()
-//)
-//
-//enum class Frequency(title: String) {
-//
-//}
